Scripts que usan librerías/requirements.txt
Pillow
opencv-python
numpy
pyautogui
pynput
pyinstaller
matplotlib
aspose-words
